Fixed IPlayerData Level type. Improved AdminPanel.

+++ Project/Assets/Sparta/SocialPoint/Social/IPlayerData.cs
@@ -18,6 +18,6 @@
         /// <summary>
         /// Player Level
         /// </summary>
-        int Level { get; }
+        long Level { get; }
     }
 }
